move interfaces to types
